updated DB logic for big Ints, fixed distance calc bug & separated logic to make easier for testing

+++ backend/pathfinder-api/create.sql
@@ -9,7 +9,7 @@ CREATE TABLE prizes (
     contract_address TEXT,
     name TEXT,
     symbol TEXT,
-    amount BIGINT,
+    amount NUMERIC,
     expires BIGINT,
     active BOOLEAN
 );
